The Big IO Update
All basic output functionality to LCD and LEDs has been completed, with some integrated into the logic of the game. That is to say, the output for guess results has been completed.

The GeneralIO source has all of the IO functions that are used generally (IE assembly and getGPIO), LCDIO has functions specific to LCD operations, and MastermindIO wraps both of these in the hardware implementation for the coursework. All IO functions used within the Logic file should be implemented there (in MastermindIO).